Create knapsackvba
The products sold in one processing will be transported in 3 trucks. The priority is to transport more than cost-valuable products to the warehouse but not exceeding the capacity of the vehicle. In the EXCEL table, the products, prices, volumes, weights and order quantities are written. The process was done with ready data